pyproject.toml
[tool.poetry]
name = "jobsearch"
version = "0.1.0"
description = ""
authors = ["Andrey <andrew.shvv@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9.17"
aiolimiter = "1.1.0"
aiomisc = {git = "https://git@github.com/aiokitchen/aiomisc"}
aiosqlite = "0.19.0"
openai = "0.27.8"
python-dotenv = "1.0.0"
python-telegram-bot = "20.4"
pytz = "2022.7"
telethon = "1.28.5"
sentry-sdk = "^1.28.1"
posthog = "^3.0.1"
chromadb = "0.4.5"
selenium = "^4.10.0"
fake-useragent = "^1.1.3"
markdownify = "^0.11.6"
tldextract = "^3.4.4"
tiktoken = "^0.4.0"
validators = "^0.20.0"
jsonschema = "^4.18.4"
langid = "^1.1.6"
phonenumbers = "^8.13.17"
pypdf2 = "^3.0.1"
python-docx = "^0.8.11"
nltk = "^3.8.1"
webdriver-manager = "^4.0.0"
pyee = "^11.0.0"
pandas = "^2.0.3"


[tool.poetry.group.dev.dependencies]
send2trash = "^1.8.2"
invoke = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
